1140 - twitter - minor refactoring
