+++ api/src/main/java/com/stormpath/sdk/provider/TwitterRequestFactory.java
@@ -3,7 +3,7 @@ package com.stormpath.sdk.provider;
 /**
  * Twitter-specific {@link ProviderRequestFactory} interface.
  *
- * @since 1.0.beta
+ * @since 1.3.0
  */
 public interface TwitterRequestFactory extends ProviderRequestFactory<TwitterAccountRequestBuilder, TwitterCreateProviderRequestBuilder> {
 }
