+++ impl/src/main/java/com/stormpath/sdk/impl/provider/DefaultTwitterCreateProviderRequestBuilder.java
@@ -12,7 +12,7 @@ public class DefaultTwitterCreateProviderRequestBuilder extends AbstractCreatePr
 
     @Override
     protected String getConcreteProviderId() {
-        return IdentityProviderType.FACEBOOK.getNameKey();
+        return IdentityProviderType.TWITTER.getNameKey();
     }
 
     @Override
